+++ packages/contracts/contracts/voting/token/TokenVotingSetup.sol
@@ -104,7 +104,7 @@ contract TokenVotingSetup is PluginSetup {
         returns (
             address plugin,
             address[] memory helpers,
-            PermissionLib.ItemMultiTarget[] memory permissions
+            PermissionLib.MultiTargetPermission[] memory permissions
         )
     {
         IDAO dao = IDAO(_dao);
@@ -189,11 +189,11 @@ contract TokenVotingSetup is PluginSetup {
         );
 
         // Prepare permissions
-        permissions = new PermissionLib.ItemMultiTarget[](tokenSettings.addr != address(0) ? 3 : 4);
+        permissions = new PermissionLib.MultiTargetPermission[](tokenSettings.addr != address(0) ? 3 : 4);
 
         // Set plugin permissions to be granted.
         // Grant the list of prmissions of the plugin to the DAO.
-        permissions[0] = PermissionLib.ItemMultiTarget(
+        permissions[0] = PermissionLib.MultiTargetPermission(
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
@@ -201,7 +201,7 @@ contract TokenVotingSetup is PluginSetup {
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
-        permissions[1] = PermissionLib.ItemMultiTarget(
+        permissions[1] = PermissionLib.MultiTargetPermission(
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
@@ -210,7 +210,7 @@ contract TokenVotingSetup is PluginSetup {
         );
 
         // Grant `EXECUTE_PERMISSION` of the DAO to the plugin.
-        permissions[2] = PermissionLib.ItemMultiTarget(
+        permissions[2] = PermissionLib.MultiTargetPermission(
             PermissionLib.Operation.Grant,
             _dao,
             plugin,
@@ -221,7 +221,7 @@ contract TokenVotingSetup is PluginSetup {
         if (tokenSettings.addr == address(0)) {
             bytes32 tokenMintPermission = GovernanceERC20(token).MINT_PERMISSION_ID();
 
-            permissions[3] = PermissionLib.ItemMultiTarget(
+            permissions[3] = PermissionLib.MultiTargetPermission(
                 PermissionLib.Operation.Grant,
                 token,
                 _dao,
@@ -242,7 +242,7 @@ contract TokenVotingSetup is PluginSetup {
         address _plugin,
         address[] calldata _helpers,
         bytes calldata
-    ) external view returns (PermissionLib.ItemMultiTarget[] memory permissions) {
+    ) external view returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         // Prepare permissions.
         uint256 helperLength = _helpers.length;
         if (helperLength != 1) {
@@ -261,10 +261,10 @@ contract TokenVotingSetup is PluginSetup {
         // Then it's GovernanceERC20.
         bool isGovernanceERC20 = supportedIds[0] && supportedIds[1] && !supportedIds[2];
 
-        permissions = new PermissionLib.ItemMultiTarget[](isGovernanceERC20 ? 4 : 3);
+        permissions = new PermissionLib.MultiTargetPermission[](isGovernanceERC20 ? 4 : 3);
 
         // Set permissions to be Revoked.
-        permissions[0] = PermissionLib.ItemMultiTarget(
+        permissions[0] = PermissionLib.MultiTargetPermission(
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
@@ -272,7 +272,7 @@ contract TokenVotingSetup is PluginSetup {
             tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
-        permissions[1] = PermissionLib.ItemMultiTarget(
+        permissions[1] = PermissionLib.MultiTargetPermission(
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
@@ -280,7 +280,7 @@ contract TokenVotingSetup is PluginSetup {
             tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
-        permissions[2] = PermissionLib.ItemMultiTarget(
+        permissions[2] = PermissionLib.MultiTargetPermission(
             PermissionLib.Operation.Revoke,
             _dao,
             _plugin,
@@ -294,7 +294,7 @@ contract TokenVotingSetup is PluginSetup {
         // calls dao for the permissions, we might decide to include the below always as it will be
         // more gas less than checking isGovernanceERC20..
         if (isGovernanceERC20) {
-            permissions[3] = PermissionLib.ItemMultiTarget(
+            permissions[3] = PermissionLib.MultiTargetPermission(
                 PermissionLib.Operation.Revoke,
                 token,
                 _dao,