+++ packages/contracts/CHANGELOG.md
@@ -24,12 +24,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `_canVote` logic to revert for votes being cast with `VoteOption.None`.
 - Changed `callId` in `DAO.execute()` from `uint256` to `bytes32`.
 - Changed all occurences of `oracle` to `condition`.
+- Renamed `ItemSingleTarget` and `ItemMultiTarget` in `PermissionLib` to `SingleTargetPermission` and `MultiTargetPermission`.
+- Renamed `bulkOnSingleTarget()` and `bulkOnMultiTarget()` in `PermissionManager` to `applySingleTargetPermissions()` and `applyMultiTargetPermissions()`
 
 ### Removed
 
 - Removes `open` parameter from Multisig proposals.
 - Removes `Tally` struct as well as `addressListLength` and moves `approvals` in `Proposal`.
 - Removes `freeze` functionality.
+- removes EIP-165 implementation from `ProposalBase`.
 
 ## v0.4.0-alpha